Add channel content view
