+++ frontend/index.html
@@ -1,6 +1,9 @@
 <!DOCTYPE html>
 <html lang="en">
     <head>
+        <meta charset="utf-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+
         <link rel="stylesheet" href="/static/css/main.css" />
         <link rel="stylesheet" href="/static/themes/default.css" />
     </head>
