+++ frontend/src/channel_views/mod.rs
@@ -1,2 +1,3 @@
 pub mod channel;
+pub mod channel_content;
 pub mod channel_message_error;
