+++ frontend/static/css/app.css
@@ -1,15 +1,23 @@
+.app-container {
+    width: 100%;
+    height: 100%;
+}
+
 .app {
     display: flex;
     justify-content: stretch;
     align-items: start;
+    height: 100%;
 }
 
 .app-navigator {
     width: 14em;
+    height: 100%;
 }
 
 .app-content {
     width: calc(100% - 14em);
+    height: 100%;
 }
 
 .user-profile-container {
