+++ frontend/static/css/channel_views/channel.css
@@ -1,3 +1,27 @@
+.channel-container {
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+}
+
+.channel-content {
+    overflow-x: hidden;
+    overflow-y: auto;
+    height: 100%;
+}
+
+.channel-content-inner {
+    min-height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-end;
+}
+
+.channel-send-button-container {
+    height: 3em;
+    background-color: var(--background);
+}
+
 .channel-message {
     display: flex;
     justify-content: left;
